[AA.1.4] Add optional functionality through incomplete instantiation
The commit provides realization of "optional functionality through
incomplete instantiation" feature presented in chapter 1.8 of the book.

The main idea is that classes parametrized with a template parameters
compile only what they use. In the example `WidgetManager` can be used
with `CreatorPolices` that do not support prototypes as long as
`WidgetManager::SwitchPrototype` method is not called.