[AA.2.3] Add Type-to-Type mapping added
The following commit provides implementation of Type-To-Type
mapping (book chapter 2.5). The mapping use example shows how to bypass
the language lack of support for function partial specialization.

In the example there is a generaic factory method with specialization for
the . The widget is created by the specialization so that
the second construction argument is set to some default value.