[AA.2.5] Add super-sub class detection
The following commit provides an implementation of compile-time
mechanism for discrimination between super and sub classes.
The mechanism is based on combination of function overloading and
sizeof function. Due to type deduction rules provided by the
language, function overload is selected - either one that
implicitly casts to a type or to elipsis. Sizeof discriminates
by returned type which overload was used.