[AA-3.7] Add Partially Ordering typelist functionality
The following commit provides an implementation of the Partially Ordering of
a TypeList (the book - chapter 3.12). The functionality allows to find the
most derived type in a single-branched inheritence hierarchy and put it
on the front of the list. The algorithm does not discriminate between
hierarchy's siblings. In case another sibling exists, the last sibling's
branch will be evaluated.