[AA-3.10] Add Generating Linear Hierarchies
The following commit provides an implementation on GenLinearHierarchy (the book chapter 3.13).
The Linear Hierarchy provides a solution to a size problem of GenScatterHierarchy. The problem
for Scatter Hierarchy is that for each branch in a hierarchy there compiler will add new virtual
table hence bloating the size of the Scatter object. The Linear Hierarchy solves the issuse
ensuring there is single virtual table per object.